+++ doc/conf.py
@@ -83,6 +83,8 @@ intersphinx_mapping = {
     "hyperspy": ("https://hyperspy.org/hyperspy-doc/current", None),
     "h5py": ("https://docs.h5py.org/en/stable", None),
     "imageio": ("https://imageio.readthedocs.io/en/stable", None),
+    "IPython": ("https://ipython.readthedocs.io/en/stable", None),
+    "ipywidgets": ("https://ipywidgets.readthedocs.io/en/stable", None),
     "matplotlib": ("https://matplotlib.org/stable", None),
     "numba": ("https://numba.readthedocs.io/en/latest", None),
     "numpy": ("https://numpy.org/doc/stable", None),