+++ CHANGELOG.rst
@@ -27,6 +27,8 @@ Added
   (`#785 <https://github.com/pyxem/kikuchipy/pull/785>`_)
 - Dependency on typing-extensions for type hints only available from Python >= 3.11.
   (`#785 <https://github.com/pyxem/kikuchipy/pull/785>`_)
+- Optional dependency on ipywidgets and IPython for IPython widgets.
+  (`#? <https://github.com/pyxem/kikuchipy/pull/?>`_)
 
 Changed
 -------